requirements.txt
fastapi==0.128.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.45
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic-core==2.41.5
pydantic-settings==2.12.0
jinja2==3.1.3
python-multipart==0.0.6
alembic==1.13.1
python-dotenv==1.2.1